app/server/requirements.txt
fastapi==0.86.0
pydantic==1.10.2
uvicorn==0.19
python-dotenv==0.21.0
SQLAlchemy==1.4.41
asyncpg==0.27.0
uuid==1.30
sqlmodel==0.0.8
python-jose==3.3.0
cryptography==38.0.1
passlib==1.7.4
bcrypt==4.0.1
sqladmin==0.7.0
httpx==0.23.1
pytest==7.2.0
pytest-asyncio==0.20.2
pytest-mock==3.10.0
pytest-httpx==0.21.2
